+++ dearpet/src/component/ShippingInfo.js
@@ -1,11 +1,46 @@
-import React, {useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import { Box, TextField, Typography, Button } from '@mui/material';
 import AddressModal from './AddressModal';
 
-const ShippingInfo = ( { address }) => {
+const ShippingInfo = ( { onAddressChange }) => {
     const [open, setOpen] = useState(false);
+    const [address, setAddress] = useState('');
     const handleOpen = () => setOpen(true);
     const handleClose = () => setOpen(false);
+
+    const fetchAddress = async () => {
+        try {
+            const token = localStorage.getItem('token');
+            const response = await fetch('http://localhost:8080/api/profile/addresses', {
+                method: 'GET',
+                headers: {
+                  Authorization: `Bearer ${token}`,
+                },
+            });
+
+            if (!response.ok) {
+                throw new Error('주소 목록을 가져오는 데 실패했습니다.');
+            }
+    
+            const data = await response.json();
+            const defaultAddr = data.find((addr) => addr.defaultAddress === true);
+            if (defaultAddr) {
+                setAddress(defaultAddr.address);
+                onAddressChange(defaultAddr.address);
+            }
+          } catch (error) {
+            console.error('주소 목록을 가져오는 데 실패했습니다:', error);
+          }
+    }
+
+    useEffect(() => {
+        fetchAddress();
+    }, []);
+    
+    const handleAddressChange = (newAddress) => {
+        setAddress(newAddress);
+        onAddressChange(newAddress);
+    };
 
     return (
         <Box mt={4}>
@@ -44,7 +79,7 @@ const ShippingInfo = ( { address }) => {
                     }}>
                     배송지 변경
                 </Button>
-                <AddressModal open={open} onClose={handleClose} />
+                <AddressModal open={open} onClose={handleClose} onAddressChange={handleAddressChange}/>
             </Box>
             {/* 배송주소에 선택된 주소 표시 */}
             <Box mb={2} sx={{ display: 'flex', alignItems: 'center' }}>
